Move "update-feature" view to attributes app
